fix responsive card 2
