+++ src/TelaEspecifica/index.js
@@ -23,16 +23,6 @@ function TelaEspecifica() {
         <div>
           <DashInternoTela />
         </div>
-        <div className={styles. telaContainer}>
-        {/* <div className={styles.chartContainer}>
-          <ContainerChart props={<ChartBar/>}/>
-          <ContainerChart props={<ChartDonut title="Quantidade de Votos" data={chartData} />}/>
-          <ContainerChart props={<ChartSemanal/>}/>
-        </div>
-        <div className={styles.listContainer}>
-          <RegistroVotos />
-        </div> */}
-        </div>
       </div>
     </TurmaProvider>
   );
