+++ src/components/VtInternaCards/index.js
@@ -3,19 +3,19 @@ import { Carousel } from "primereact/carousel";
 import { PieChart, Pie, Cell } from "recharts";
 import styles from "./VtInternaCards.module.css";
 import { TurmaContext } from "../../context/TurmaContext";
-
-import ChartSemanal from '../ChartSemanal';
-import ChartDonut from '../ChartDonut';
-import ChartBar from '../ChartBar';
-import ContainerChart from '../ContainerChart';
+import TelaEspecifica from "../../TelaEspecifica";
+import Charts from "../../charts";
 import RegistroVotos from "../../components/ListRegistroVotos";
+// import ChartSemanal from '../ChartSemanal';
+// import ChartDonut from '../ChartDonut';
+// import ChartBar from '../ChartBar';
+// import ContainerChart from '../ContainerChart';
+// import RegistroVotos from "../../components/ListRegistroVotos";
 
-
-
-const chartData = [
-  { name: "Votos Confirmados", value: 50 },
-  { name: "Não Votaram", value: 50 },
-];
+// const chartData = [
+//   { name: "Votos Confirmados", value: 50 },
+//   { name: "Não Votaram", value: 50 },
+// ];
 
 const COLORS = ["#ff0000", "#000000"];
 
@@ -212,19 +212,16 @@ const VtInternaCards = () => {
               ))}
             </div>
           </div>
-
-          <div className={styles.chartContainer}>
-            <ContainerChart props={<ChartBar />} />
-            <ContainerChart
-              props={
-                <ChartDonut title="Quantidade de Votos" data={chartData} />
-              }
-            />
-            <ContainerChart props={<ChartSemanal />} />
+          <div className={styles.telaContainer}>
+            <div className={styles.charts}>
+              <Charts />
+            </div>
           </div>
           <div className={styles.listContainer}>
-          <RegistroVotos />
-        </div>
+            <div className={styles.registro}>
+              <RegistroVotos />
+            </div>
+          </div>
         </>
       )}
     </>
